+++ app/src/main/java/com/example/practica4/data/Bluetooth.kt
@@ -5,25 +5,32 @@ import android.app.Activity
 import android.app.AlertDialog
 import android.bluetooth.BluetoothAdapter
 import android.bluetooth.BluetoothManager
+import android.bluetooth.BluetoothSocket
 import android.bluetooth.le.ScanCallback
 import android.bluetooth.le.ScanResult
 import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Build
+import android.util.Log
 import androidx.activity.result.ActivityResultLauncher
 import androidx.compose.runtime.mutableStateListOf
 import androidx.core.app.ActivityCompat
 import androidx.core.content.ContextCompat
+import java.io.IOException
+import java.util.UUID
 
-class BluetoothRepository(context: Context) {
+class BluetoothRepository {
     val devices = mutableStateListOf<String>()
     val pairedDevices = mutableStateListOf<String>()
     private var scanCallback: ScanCallback? = null
+    private var bluetoothSocket: BluetoothSocket? = null
+    private val sppUUID = UUID.fromString("00001101-0000-1000-8000-00805F9B34FB")
+
     companion object {
         private const val BLUETOOTH_PERMISSION_REQUEST_CODE = 100
     }
 
-    fun getBluetoothAdapter(context: Context): BluetoothAdapter? {
+    private fun getBluetoothAdapter(context: Context): BluetoothAdapter? {
         val bluetoothManager = context.getSystemService(BluetoothManager::class.java)
         return bluetoothManager?.adapter
     }
@@ -139,4 +146,104 @@ class BluetoothRepository(context: Context) {
             }
         }
     }
+
+    fun connectToDeviceByMac(context: Context, macAddress: String): BluetoothSocket {
+        if (!hasBluetoothPermissions(context)) {
+            throw SecurityException("Permisos de Bluetooth no concedidos")
+        }
+
+        val bluetoothAdapter = getBluetoothAdapter(context)
+            ?: throw IllegalStateException("Bluetooth no disponible")
+
+        if (!bluetoothAdapter.isEnabled) {
+            throw IllegalStateException("Bluetooth está desactivado")
+        }
+
+        val device = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+            if (ContextCompat.checkSelfPermission(context, Manifest.permission.BLUETOOTH_CONNECT) == PackageManager.PERMISSION_GRANTED) {
+                bluetoothAdapter.getRemoteDevice(macAddress)
+            } else {
+                throw SecurityException("Permiso BLUETOOTH_CONNECT requerido")
+            }
+        } else {
+            bluetoothAdapter.getRemoteDevice(macAddress)
+        }
+
+        val socket = device.createRfcommSocketToServiceRecord(sppUUID)
+        bluetoothAdapter.cancelDiscovery()
+        socket.connect()
+        bluetoothSocket = socket
+        return socket
+    }
+
+    fun disconnect() {
+        try {
+            bluetoothSocket?.close()
+        } catch (e: IOException) {
+            Log.e("BluetoothRepository", "Error al cerrar el socket: ${e.message}")
+        }
+        bluetoothSocket = null
+    }
+
+    fun sendCommand(command: String, onResponse: (String) -> Unit) {
+        bluetoothSocket?.let { socket ->
+            Thread {
+                try {
+                    socket.outputStream.write(command.toByteArray())
+                    socket.outputStream.flush()
+
+                    val startTime = System.currentTimeMillis()
+                    val buffer = ByteArray(1024)
+                    var response = ""
+
+                    while (System.currentTimeMillis() - startTime < 2000) {
+                        if (socket.inputStream.available() > 0) {
+                            val bytesRead = socket.inputStream.read(buffer)
+                            response += String(buffer, 0, bytesRead)
+                            if (response.contains('\n')) break
+                        }
+                        Thread.sleep(50)
+                    }
+
+                    if (response.isNotEmpty()) {
+                        onResponse(response.trim())
+                    } else {
+                        Log.w("BluetoothRepository", "No hubo respuesta del Arduino")
+                    }
+
+                } catch (e: IOException) {
+                    Log.e("BluetoothRepository", "Error al enviar comando: ${e.message}")
+                }
+            }.start()
+        }
+    }
+
+    fun requestPermissions(
+        activity: Activity,
+        scanPermissionLauncher: ActivityResultLauncher<String>,
+        connectPermissionLauncher: ActivityResultLauncher<String>
+    ) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+            if (ActivityCompat.shouldShowRequestPermissionRationale(activity, Manifest.permission.BLUETOOTH_SCAN)) {
+                AlertDialog.Builder(activity)
+                    .setTitle("Permisos necesarios")
+                    .setMessage("La aplicación necesita permisos de Bluetooth para escanear y conectar dispositivos")
+                    .setPositiveButton("Entendido") { _, _ ->
+                        scanPermissionLauncher.launch(Manifest.permission.BLUETOOTH_SCAN)
+                        connectPermissionLauncher.launch(Manifest.permission.BLUETOOTH_CONNECT)
+                    }
+                    .show()
+            } else {
+                scanPermissionLauncher.launch(Manifest.permission.BLUETOOTH_SCAN)
+                connectPermissionLauncher.launch(Manifest.permission.BLUETOOTH_CONNECT)
+            }
+        } else {
+            ActivityCompat.requestPermissions(
+                activity,
+                arrayOf(Manifest.permission.ACCESS_FINE_LOCATION),
+                100
+            )
+        }
+    }
+
 }
